+++ exts/omni.isaac.lab_quadruped_tasks/omni/isaac/lab_quadruped_tasks/robots/__init__.py
@@ -3,4 +3,5 @@ License: MIT License
 Copyright (c) 2024, Felipe Mohr Santos
 """
 
+from . import spot  # noqa: F401, F403
 from . import unitree_go2  # noqa: F401, F403