+++ exts/omni.isaac.lab_quadruped_tasks/docs/CHANGELOG.rst
@@ -8,6 +8,7 @@ Added
 ^^^^^
 
 * Create `cfg`, `robot` and `agent` folders
+* Create tasks for Spot robot from Boston Dynamics
 
 Changed
 ^^^^^